Reduce duplicate code in camera control (example)
