+++ src/main/drivers/camera_control.c
@@ -1,19 +1,20 @@
 /*
- * This file is part of Cleanflight and Betaflight.
+ * This file is part of Betaflight.
  *
- * Cleanflight and Betaflight are free software. You can redistribute
- * this software and/or modify this software under the terms of the
- * GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option)
- * any later version.
+ * Betaflight is free software. You can redistribute this software
+ * and/or modify this software under the terms of the GNU General
+ * Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later
+ * version.
  *
- * Cleanflight and Betaflight are distributed in the hope that they
- * will be useful, but WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * Betaflight is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
  * See the GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this software.
+ * You should have received a copy of the GNU General Public
+ * License along with this software.
  *
  * If not, see <http://www.gnu.org/licenses/>.
  */
@@ -29,29 +30,16 @@
 #include <math.h>
 
 #include "drivers/camera_control_impl.h"
-#include "drivers/rcc.h"
 #include "drivers/io.h"
-#include "drivers/nvic.h"
 #include "drivers/pwm_output.h"
 #include "drivers/time.h"
 #include "pg/pg_ids.h"
-
-#define CAMERA_CONTROL_PWM_RESOLUTION   128
-#define CAMERA_CONTROL_SOFT_PWM_RESOLUTION 448
 
 #ifdef CURRENT_TARGET_CPU_VOLTAGE
 #define ADC_VOLTAGE CURRENT_TARGET_CPU_VOLTAGE
 #else
 #define ADC_VOLTAGE 3.3f
 #endif
-
-#if !defined(STM32F411xE) && !defined(STM32F7) && !defined(STM32H7) && !defined(STM32G4)
-#define CAMERA_CONTROL_SOFTWARE_PWM_AVAILABLE
-#include "build/atomic.h"
-#endif
-
-#define CAMERA_CONTROL_HARDWARE_PWM_AVAILABLE
-#include "drivers/timer.h"
 
 #ifdef USE_OSD
 #include "osd/osd.h"
@@ -85,7 +73,7 @@ static struct {
 static uint32_t endTimeMillis;
 
 #ifdef CAMERA_CONTROL_SOFTWARE_PWM_AVAILABLE
-static void cameraControlHi(void)
+void cameraControlHi(void)
 {
     if (cameraControlRuntime.inverted) {
         IOLo(cameraControlRuntime.io);
@@ -94,26 +82,13 @@ static void cameraControlHi(void)
     }
 }
 
-static void cameraControlLo(void)
+void cameraControlLo(void)
 {
     if (cameraControlRuntime.inverted) {
         IOHi(cameraControlRuntime.io);
     } else {
         IOLo(cameraControlRuntime.io);
     }
-}
-
-void TIM6_DAC_IRQHandler(void)
-{
-    cameraControlHi();
-    TIM6->SR = 0;
-}
-
-void TIM7_IRQHandler(void)
-{
-    cameraControlLo();
-
-    TIM7->SR = 0;
 }
 #endif
 
@@ -136,7 +111,7 @@ void cameraControlInit(void)
 
         IOConfigGPIOAF(cameraControlRuntime.io, IOCFG_AF_PP, timerHardware->alternateFunction);
 
-        pwmOutConfig(&cameraControlRuntime.channel, timerHardware, timerClock(TIM6), CAMERA_CONTROL_PWM_RESOLUTION, 0, cameraControlRuntime.inverted);
+        cameraControlHardwarePwmInit(&cameraControlRuntime.channel, timerHardware, cameraControlRuntime.inverted);
 
         cameraControlRuntime.period = CAMERA_CONTROL_PWM_RESOLUTION;
         *cameraControlRuntime.channel.ccr = cameraControlRuntime.period;
@@ -151,19 +126,7 @@ void cameraControlInit(void)
         cameraControlRuntime.period = CAMERA_CONTROL_SOFT_PWM_RESOLUTION;
         cameraControlRuntime.enabled = true;
 
-
-        NVIC_InitTypeDef nvicTIM6 = {
-            TIM6_DAC_IRQn, NVIC_PRIORITY_BASE(NVIC_PRIO_TIMER), NVIC_PRIORITY_SUB(NVIC_PRIO_TIMER), ENABLE
-        };
-        NVIC_Init(&nvicTIM6);
-        NVIC_InitTypeDef nvicTIM7 = {
-            TIM7_IRQn, NVIC_PRIORITY_BASE(NVIC_PRIO_TIMER), NVIC_PRIORITY_SUB(NVIC_PRIO_TIMER), ENABLE
-        };
-        NVIC_Init(&nvicTIM7);
-
-        RCC->APB1ENR |= RCC_APB1Periph_TIM6 | RCC_APB1Periph_TIM7;
-        TIM6->PSC = 0;
-        TIM7->PSC = 0;
+        cameraControlSoftwarePwmInit();
 #endif
     } else if (CAMERA_CONTROL_MODE_DAC == cameraControlConfig()->mode) {
         // @todo not yet implemented
@@ -183,13 +146,14 @@ void cameraControlProcess(uint32_t currentTimeUs)
     }
 }
 
+#if defined(CAMERA_CONTROL_HARDWARE_PWM_AVAILABLE) || defined(CAMERA_CONTROL_SOFTWARE_PWM_AVAILABLE)
+
 static float calculateKeyPressVoltage(const cameraControlKey_e key)
 {
     const int buttonResistance = cameraControlConfig()->buttonResistanceValues[key] * 100;
     return 1.0e-2f * cameraControlConfig()->refVoltage * buttonResistance / (100 * cameraControlConfig()->internalResistance + buttonResistance);
 }
 
-#if defined(CAMERA_CONTROL_HARDWARE_PWM_AVAILABLE) || defined(CAMERA_CONTROL_SOFTWARE_PWM_AVAILABLE)
 static float calculatePWMDutyCycle(const cameraControlKey_e key)
 {
     const float voltage = calculateKeyPressVoltage(key);
@@ -231,21 +195,7 @@ void cameraControlKeyPress(cameraControlKey_e key, uint32_t holdDurationMs)
             delay(cameraControlConfig()->keyDelayMs + holdDurationMs);
             cameraControlHi();
         } else {
-            TIM6->CNT = hiTime;
-            TIM6->ARR = cameraControlRuntime.period;
-
-            TIM7->CNT = 0;
-            TIM7->ARR = cameraControlRuntime.period;
-
-            // Start two timers as simultaneously as possible
-            ATOMIC_BLOCK(NVIC_PRIO_TIMER) {
-                TIM6->CR1 = TIM_CR1_CEN;
-                TIM7->CR1 = TIM_CR1_CEN;
-            }
-
-            // Enable interrupt generation
-            TIM6->DIER = TIM_IT_Update;
-            TIM7->DIER = TIM_IT_Update;
+            cameraControlSoftwarePwmEnable(hiTime, cameraControlRuntime.period);
 
             const uint32_t endTime = millis() + cameraControlConfig()->keyDelayMs + holdDurationMs;
 
@@ -253,10 +203,7 @@ void cameraControlKeyPress(cameraControlKey_e key, uint32_t holdDurationMs)
             while (millis() < endTime);
 
             // Disable timers and interrupt generation
-            TIM6->CR1 &= ~TIM_CR1_CEN;
-            TIM7->CR1 &= ~TIM_CR1_CEN;
-            TIM6->DIER = 0;
-            TIM7->DIER = 0;
+            cameraControlSoftwarePwmDisable();
 
             // Reset to idle state
             IOHi(cameraControlRuntime.io);
@@ -267,4 +214,4 @@ void cameraControlKeyPress(cameraControlKey_e key, uint32_t holdDurationMs)
     }
 }
 
-#endif
+#endif // USE_CAMERA_CONTROL
