+++ src/platform/STM32/mk/STM32F7.mk
@@ -169,7 +169,7 @@ MCU_COMMON_SRC = \
             timer_hal.c \
             timer_stm32f7xx.c \
             transponder_ir_io_hal.c \
-            camera_control.c \
+            camera_control_stm32.c \
             startup/system_stm32f7xx.c
 
 MCU_EXCLUDES = \
